requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
feedparser
PyYAML
